VirtualMemoryDomain
