+++ Source/Libraries/CorruptCore/CorruptCore.csproj
@@ -100,7 +100,8 @@
     <Compile Include="EventWarlock\WarlockAction.cs" />
     <Compile Include="EventWarlock\WarlockActions\WarlockActionEcho.cs" />
     <Compile Include="EventWarlock\WarlockConditions\FirstEqualsSecond.cs" />
-    <Compile Include="MemoryInterface\MemoryInterface.cs" />
+    <Compile Include="MemoryDomains\MemoryInterface.cs" />
+    <Compile Include="MemoryDomains\VirtualMemoryDomain.cs" />
     <Compile Include="IRealTime.cs" />
     <Compile Include="ListFilters.cs" />
     <Compile Include="MemoryBanks.cs" />
