+++ Source/Libraries/CorruptCore/MemoryDomains/MemoryInterface.cs
